hacky pen color state
